class 7
worked on our second project to make something that can help people during natural disastors
I made a fire alarm that detects spikes in brightness, temperature, and humidity and alerts through an LCD screen, 2 RGB LEDs, and a sprinkler (a servo). It also includes a "drill" mode to act as a fire drill of sorts.
This was our final assignment before starting on our final project as a group.